packages/entropretty-testbench/requirements.txt
numpy
opencv-python
scikit-image
scikit-learn
Pillow
ImageHash
tensorflow
matplotlib